+++ yesod-raml-mock/ChangeLog.md
@@ -1,3 +1,7 @@
+## 0.1.1
+
+* Add mockSrcFromRaml to print mock-handler
+
 ## 0.1.0
 
 * First Release